Clean criteria based files
