+++ src/components/filter/criteria-based/criteria-based-filter-form.tsx
@@ -16,7 +16,7 @@ import { FunctionComponent } from 'react';
 import {
     getCriteriaBasedFormData,
     getCriteriaBasedSchema,
-} from '../utils/criteria-based-utils';
+} from './criteria-based-filter-utils';
 
 export const criteriaBasedFilterSchema = getCriteriaBasedSchema({
     [FieldConstants.ENERGY_SOURCE]: yup.string().nullable(),
