+++ src/components/filter/utils/filter-form-utils.ts
@@ -5,63 +5,11 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 import { FieldConstants } from '../constants/field-constants';
-import yup from '../../../utils/yup-config';
-import RangeInput, {
-    DEFAULT_RANGE_VALUE,
-    getRangeInputDataForm,
-    getRangeInputSchema,
-} from '../../react-hook-form/range-input.tsx';
+import RangeInput from '../../react-hook-form/range-input';
 import { FunctionComponent } from 'react';
 import CountriesInput from '../../react-hook-form/select-inputs/countries-input.tsx';
 import SelectInput from '../../react-hook-form/select-inputs/select-input.tsx';
 
-export const getCriteriaBasedSchema = (extraFields: any) => ({
-    [FieldConstants.CRITERIA_BASED]: yup.object().shape({
-        [FieldConstants.COUNTRIES]: yup.array().of(yup.string()),
-        [FieldConstants.COUNTRIES_1]: yup.array().of(yup.string()),
-        [FieldConstants.COUNTRIES_2]: yup.array().of(yup.string()),
-        ...getRangeInputSchema(FieldConstants.NOMINAL_VOLTAGE),
-        ...getRangeInputSchema(FieldConstants.NOMINAL_VOLTAGE_1),
-        ...getRangeInputSchema(FieldConstants.NOMINAL_VOLTAGE_2),
-        ...getRangeInputSchema(FieldConstants.NOMINAL_VOLTAGE_3),
-        ...extraFields,
-    }),
-});
-
-export const getCriteriaBasedFormData = (
-    criteriaValues: any,
-    extraFields: any
-) => ({
-    [FieldConstants.CRITERIA_BASED]: {
-        [FieldConstants.COUNTRIES]:
-            criteriaValues?.[FieldConstants.COUNTRIES] ?? [],
-        [FieldConstants.COUNTRIES_1]:
-            criteriaValues?.[FieldConstants.COUNTRIES_1] ?? [],
-        [FieldConstants.COUNTRIES_2]:
-            criteriaValues?.[FieldConstants.COUNTRIES_2] ?? [],
-        ...getRangeInputDataForm(
-            FieldConstants.NOMINAL_VOLTAGE,
-            criteriaValues?.[FieldConstants.NOMINAL_VOLTAGE] ??
-                DEFAULT_RANGE_VALUE
-        ),
-        ...getRangeInputDataForm(
-            FieldConstants.NOMINAL_VOLTAGE_1,
-            criteriaValues?.[FieldConstants.NOMINAL_VOLTAGE_1] ??
-                DEFAULT_RANGE_VALUE
-        ),
-        ...getRangeInputDataForm(
-            FieldConstants.NOMINAL_VOLTAGE_2,
-            criteriaValues?.[FieldConstants.NOMINAL_VOLTAGE_2] ??
-                DEFAULT_RANGE_VALUE
-        ),
-        ...getRangeInputDataForm(
-            FieldConstants.NOMINAL_VOLTAGE_3,
-            criteriaValues?.[FieldConstants.NOMINAL_VOLTAGE_3] ??
-                DEFAULT_RANGE_VALUE
-        ),
-        ...extraFields,
-    },
-});
 const countries = {
     renderer: CountriesInput,
     props: {
@@ -126,59 +74,57 @@ const energySource = {
         ],
     },
 };
-type CriteriaFormField = {
+type FilterFormField = {
     renderer: FunctionComponent<any>;
     props: {
         label: string;
         name: string;
     };
 };
-type CriteriaFormEquipment = {
+type FilterFormEquipment = {
     id: string;
     label: string;
-    fields: CriteriaFormField[];
+    fields: FilterFormField[];
 };
-export const CONTINGENCY_LIST_EQUIPMENTS: Record<
-    string,
-    CriteriaFormEquipment
-> = {
-    LINE: {
-        id: 'LINE',
-        label: 'Lines',
-        fields: [countries1, countries2, nominalVoltage1, nominalVoltage2],
-    },
-    TWO_WINDINGS_TRANSFORMER: {
-        id: 'TWO_WINDINGS_TRANSFORMER',
-        label: 'TwoWindingsTransformers',
-        fields: [countries, nominalVoltage1, nominalVoltage2],
-    },
-    GENERATOR: {
-        id: 'GENERATOR',
-        label: 'Generators',
-        fields: [countries, nominalVoltage],
-    },
-    SHUNT_COMPENSATOR: {
-        id: 'SHUNT_COMPENSATOR',
-        label: 'ShuntCompensators',
-        fields: [countries, nominalVoltage],
-    },
-    HVDC_LINE: {
-        id: 'HVDC_LINE',
-        label: 'HvdcLines',
-        fields: [countries1, countries2, nominalVoltage],
-    },
-    BUSBAR_SECTION: {
-        id: 'BUSBAR_SECTION',
-        label: 'BusBarSections',
-        fields: [countries, nominalVoltage],
-    },
-    DANGLING_LINE: {
-        id: 'DANGLING_LINE',
-        label: 'DanglingLines',
-        fields: [countries, nominalVoltage],
-    },
-};
-export const FILTER_EQUIPMENTS: Record<string, CriteriaFormEquipment> = {
+export const CONTINGENCY_LIST_EQUIPMENTS: Record<string, FilterFormEquipment> =
+    {
+        LINE: {
+            id: 'LINE',
+            label: 'Lines',
+            fields: [countries1, countries2, nominalVoltage1, nominalVoltage2],
+        },
+        TWO_WINDINGS_TRANSFORMER: {
+            id: 'TWO_WINDINGS_TRANSFORMER',
+            label: 'TwoWindingsTransformers',
+            fields: [countries, nominalVoltage1, nominalVoltage2],
+        },
+        GENERATOR: {
+            id: 'GENERATOR',
+            label: 'Generators',
+            fields: [countries, nominalVoltage],
+        },
+        SHUNT_COMPENSATOR: {
+            id: 'SHUNT_COMPENSATOR',
+            label: 'ShuntCompensators',
+            fields: [countries, nominalVoltage],
+        },
+        HVDC_LINE: {
+            id: 'HVDC_LINE',
+            label: 'HvdcLines',
+            fields: [countries1, countries2, nominalVoltage],
+        },
+        BUSBAR_SECTION: {
+            id: 'BUSBAR_SECTION',
+            label: 'BusBarSections',
+            fields: [countries, nominalVoltage],
+        },
+        DANGLING_LINE: {
+            id: 'DANGLING_LINE',
+            label: 'DanglingLines',
+            fields: [countries, nominalVoltage],
+        },
+    };
+export const FILTER_EQUIPMENTS: Record<string, FilterFormEquipment> = {
     ...CONTINGENCY_LIST_EQUIPMENTS,
     THREE_WINDINGS_TRANSFORMER: {
         id: 'THREE_WINDINGS_TRANSFORMER',
